+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/WithinBlockTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2022 DiffPlug
+ * Copyright 2020-2023 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -31,7 +31,7 @@ class WithinBlockTest extends GradleIntegrationHarness {
 				"spotless {",
 				"  format 'customJava', JavaExtension, {",
 				"    target '*.java'",
-				"    googleJavaFormat('1.2')",
+				"    googleJavaFormat()",
 				"  }",
 				"}");
 		setFile("test.java").toResource("java/googlejavaformat/JavaCodeUnformatted.test");
@@ -53,7 +53,7 @@ class WithinBlockTest extends GradleIntegrationHarness {
 				"      custom 'lowercase', { str -> str.toLowerCase() }",
 				"    }",
 				"    withinBlocks 'java only', '\\n```java\\n', '\\n```\\n', JavaExtension, {",
-				"      googleJavaFormat('1.2')",
+				"      googleJavaFormat()",
 				"    }",
 				"  }",
 				"}");