+++ plugin-maven/src/test/java/com/diffplug/spotless/maven/MavenProvisionerTest.java
@@ -33,7 +33,7 @@ class MavenProvisionerTest extends MavenIntegrationHarness {
 	void testSingleDependencyIncludingTransitives() throws Exception {
 		writePomWithJavaSteps(
 				"<googleJavaFormat>",
-				"  <version>1.2</version>",
+				"  <version>1.10.0</version>",
 				"</googleJavaFormat>");
 		assertResolveDependenciesWorks();
 	}